+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-# fights CHANGELOG.md
+# Change Log
 
 ## Unreleased
+
+## 0.2.0-SNAPSHOT
+- src/fights/db.clj create tables. `persons`, `contests` and `competitions`.
+- test/fights/db_test.clj
+- src/fights/competitor.clj
+- src/fights/core.clj
+
 
 ## 0.1.0 - 2022-08-12
 - project (re)started.